+++ engine/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.22.0-dev
+* Fix READMEs [#365](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/365)
 
 ## 0.21.0
 * Rename the function `is_simulcast` to `simulcast?` in order to be compliant with elixir style guide. [#349](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/349)